refactor: separate a single testcase component in own file
